libreoffice-no gui, move update out of fileidentification - less python dep for docker
