+++ d_requirements.txt
@@ -1,0 +1,15 @@
+annotated-types==0.7.0
+click==8.3.0
+dotenv==0.9.9
+markdown-it-py==4.0.0
+mdurl==0.1.2
+pydantic==2.11.9
+pydantic-core==2.33.2
+pygfried==0.12.0
+pygments==2.19.2
+python-dotenv==1.1.1
+rich==14.1.0
+shellingham==1.5.4
+typer==0.19.2
+typing-extensions==4.15.0
+typing-inspection==0.4.1
