--- requirements.txt
@@ -1,22 +1,0 @@
-annotated-types==0.7.0
-beautifulsoup4==4.13.1
-certifi==2024.12.14
-charset-normalizer==3.4.1
-click==8.1.8
-idna==3.10
-lxml==5.3.0
-markdown-it-py==3.0.0
-mdurl==0.1.2
-pydantic==2.11.7
-pydantic-core==2.33.2
-pygfried==0.12.0
-pygments==2.19.1
-requests==2.32.3
-rich==13.9.4
-shellingham==1.5.4
-soupsieve==2.6
-typer==0.15.1
-typing-extensions==4.12.2
-typing-inspection==0.4.1
-urllib3==2.3.0
-dotenv==0.9.9
